xphan - finish place api, change orderstatus

+++ src/main/java/com/cs490/onlineshopping/admin/repository/ProductRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Sort;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends PagingAndSortingRepository<Product,Long> {
+public interface ProductRepository extends PagingAndSortingRepository<Product,Integer> {
 
     List<Product> findByVendor(Vendor vendor);
 }
